requirements.txt
# conquest.ai — Python Dependencies
#
# Install with: pip install -r requirements.txt

# --- UI Framework ---
chainlit>=1.3.0

# --- LLM & RAG Orchestration ---
llama-index>=0.11.0
llama-index-vector-stores-chroma>=0.3.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-llms-anthropic>=0.6.0

# --- Vector Database ---
chromadb>=0.5.0

# --- Document Parsers ---
pymupdf>=1.24.0          # PDF parsing (fitz)
ebooklib>=0.18           # EPUB parsing
beautifulsoup4>=4.12.0   # HTML stripping for EPUB
python-docx>=1.1.0       # DOCX parsing

# --- Embeddings (local, no API required) ---
sentence-transformers>=3.0.0

# --- Utilities ---
python-dotenv>=1.0.0     # .env file support
